+++ CHANGELOG.rst
@@ -17,13 +17,17 @@ Changes to "0.7.1-alpha"
 Features
 """"""""
 
-- Python: mpi4py support added #454
+- Python:
+
+  - mpi4py support added #454
+  - slice protocol for record component #458
 
 Bug Fixes
 """""""""
 
 - do not require write permissions to open ``Series`` read-only #395
 - loadChunk: re-enable range/extent checks for adjusted ranges #469
+- Python: stricter contiguous check for user-provided arrays #458
 
 Other
 """""